feat: Refactor admin and reports services, adding modular functions for better maintainability and clarity
